requirements.txt
streamlit
flask
PyMuPDF
google-generativeai
